Add User Interface module
