+++ Sources/PrimerSDK/Classes/User Interface/Primer/PaymentMethodComponent.swift
@@ -27,8 +27,8 @@ internal class PaymentMethodComponent: PrimerView {
             label.textColor = .white
             addSubview(label)
             addSubview(iconView)
-            configureLabel(with: method.buttonTitle, isBold: true)
-            configureIconView(icon: method.buttonImage, color: .white, isMonoColor: true)
+            configureLabel(with: method.uiModule.buttonTitle, isBold: true)
+            configureIconView(icon: method.uiModule.buttonImage, color: .white, isMonoColor: true)
             anchorLabel()
             anchorIconView(inRelationToLabel: true)
         case .paymentCard:
@@ -38,8 +38,8 @@ internal class PaymentMethodComponent: PrimerView {
             label.textColor = theme.paymentMethodButton.text.color
             addSubview(label)
             addSubview(iconView)
-            configureLabel(with: method.buttonTitle)
-            configureIconView(icon: method.buttonImage, color: theme.paymentMethodButton.text.color, isMonoColor: true)
+            configureLabel(with: method.uiModule.buttonTitle)
+            configureIconView(icon: method.uiModule.buttonImage, color: theme.paymentMethodButton.text.color, isMonoColor: true)
             anchorLabel()
             anchorIconView(inRelationToLabel: true)
         case .payPal:
@@ -47,7 +47,7 @@ internal class PaymentMethodComponent: PrimerView {
             backgroundColor = UIColor(red: 190/255, green: 228/255, blue: 254/255, alpha: 1)
             //            layer.borderColor = theme.colorTheme.disabled1.cgColor
             addSubview(iconView)
-            configureIconView(icon: method.buttonImage, color: theme.paymentMethodButton.text.color)
+            configureIconView(icon: method.uiModule.buttonImage, color: theme.paymentMethodButton.text.color)
             anchorIconView(inRelationToLabel: false)
         case .goCardlessMandate:
             layer.borderWidth = 1
@@ -56,15 +56,15 @@ internal class PaymentMethodComponent: PrimerView {
             label.textColor = theme.paymentMethodButton.text.color
             addSubview(label)
             addSubview(iconView)
-            configureLabel(with: method.buttonTitle)
-            configureIconView(icon: method.buttonImage, color: theme.paymentMethodButton.text.color, isMonoColor: true)
+            configureLabel(with: method.uiModule.buttonTitle)
+            configureIconView(icon: method.uiModule.buttonImage, color: theme.paymentMethodButton.text.color, isMonoColor: true)
             anchorLabel()
             anchorIconView(inRelationToLabel: true)
         case .klarna:
             // TODO: move Klarna color to constant or similar, should maybe be dynamic from backend?
             backgroundColor = UIColor(red: 255/255, green: 179/255, blue: 199/255, alpha: 1)
             addSubview(iconView)
-            configureIconView(icon: method.buttonImage, color: theme.paymentMethodButton.text.color)
+            configureIconView(icon: method.uiModule.buttonImage, color: theme.paymentMethodButton.text.color)
             anchorIconView(inRelationToLabel: false)
         case .apaya:
             layer.borderWidth = 1
@@ -73,8 +73,8 @@ internal class PaymentMethodComponent: PrimerView {
             label.textColor = theme.paymentMethodButton.text.color
             addSubview(label)
             addSubview(iconView)
-            configureLabel(with: method.buttonTitle)
-            configureIconView(icon: method.buttonImage, color: theme.paymentMethodButton.text.color, isMonoColor: true)
+            configureLabel(with: method.uiModule.buttonTitle)
+            configureIconView(icon: method.uiModule.buttonImage, color: theme.paymentMethodButton.text.color, isMonoColor: true)
             anchorLabel()
             anchorIconView(inRelationToLabel: true)
         default:
