+++ Sources/PrimerSDK/Classes/User Interface/TokenizationViewModels/CardFormPaymentMethodTokenizationViewModel.swift
@@ -24,6 +24,7 @@ class CardFormPaymentMethodTokenizationViewModel: PaymentMethodTokenizationViewM
     private var cardComponentsManagerTokenizationCompletion: ((PrimerPaymentMethodTokenData?, Error?) -> Void)?
     private var webViewController: SFSafariViewController?
     private var webViewCompletion: ((_ authorizationToken: String?, _ error: Error?) -> Void)?
+    private let theme: PrimerThemeProtocol = DependencyContainer.resolve()
     
     private var isCardholderNameFieldEnabled: Bool {
         if (AppState.current.apiConfiguration?.checkoutModules?.filter({ $0.type == "CARD_INFORMATION" }).first?.options as? PrimerAPIConfiguration.CheckoutModule.CardInformationOptions)?.cardHolderName == false {
